remove nprocs
The use of nprocs is confusing, because I think it is referring to tasks.
I am removing it for now, and will add back tasks when flux warrants it.
For now we assume one task per node, where each noded has some number
of cores (cores per task). Note - I also want to remove the space
in that argument, it is not great! I will do it after experiments.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>