feat: node timings (#20)
* feat: node timings

We need to make it easy to see when nodes go up and down in the
context of an experiment. This gives a cluster view of nodes,
and that permission might be too big (and also assumes that
the user can filter down to those used in the workflow) but
for our experiments, we need to have these timestamps
coupled directly with the workflow. We should be able to
easily see the node creations / deletions that are in the
same span as when a workflow is running. I am doing this
by introducing the idea of custom tracker watchers, which
are started, stopped, and then asked to save any data to some
path. We do this design because Flux will not have these
features.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* feat: add labels and print to terminal.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* add more metadata saving

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* make more resilient to error

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>