feat: save kubernetes logs.
We have been saving artifacts for everything, relying on the
application to take the burden of saving its own logging
retrieved from the registry. For experiments with gpu
selection we just need one little value, and I think
it would be easier to save all the logs instead of
using oras. This feature supports that, where the
user adds a properties -> save-path, and under that
path "logs" is created that is named by the job,
step, and pod index.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>