feat: add support for minicluster
If we really want to test scale (shrink and grow) of a job
and have it work with the cluster autoscaler, plus collecting
metrics from an HPC app, we can most easily do that with
the flux operator. This feature adds support for specifying
a minicluster property to convert the previous indexed job
into a MiniCluster. The flux operator needs to be installed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>