feat: support for custom metrics
In this example, the user is allowed to provide a custom script
that will be used against the log, and it needs to return a dictionary
of values (the custom metrics). These are passed back to the manager
from the state machine step and can influence workflow behavior (e.g.,
stop early, grow, or shrink.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>